Broke down summary_services.py into summary_services.py and summary_tile_services.py
