+++ assets/i18n/en.json
@@ -180,7 +180,7 @@
   "I18N_PLAYER_REPORT_MODAL_BODY_ADDITIONAL_DETAILS": "Please provide additional details for the moderators:",
   "I18N_PLAYER_REPORT_MODAL_BODY_HEADER": "What is the issue?",
   "I18N_PLAYER_REPORT_MODAL_BODY_OTHER": "Other",
-  "I18N_PLAYER_REPORT_MODAL_BODY_POOR_EXPERIENCE": "Poor learning experience", 
+  "I18N_PLAYER_REPORT_MODAL_BODY_POOR_EXPERIENCE": "Poor learning experience",
   "I18N_PLAYER_REPORT_MODAL_FOOTER_CANCEL": "Cancel",
   "I18N_PLAYER_REPORT_MODAL_FOOTER_SUBMIT": "Submit",
   "I18N_PLAYER_REPORT_MODAL_HEADER": "Report this exploration",
@@ -265,16 +265,14 @@
   "I18N_SIGNUP_USERNAME_EXPLANATION": "Your username will be shown next to your contributions.",
   "I18N_SIGNUP_WAIVER_OBJECTIVE": "The waiver of the attribution (BY) requirement means that, if someone reuses this work, they are not required to attribute the authors. However, all of your individual contributions to explorations will be available on this site in the exploration change log, and people who reuse the exploration are encouraged (but not required) to include a link to this page.",
   "I18N_SIGNUP_WHY_LICENSE": "Why CC-BY-SA?",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Explorations are interactive lessons that provide more immersive experiences than static videos or text, helping users learn by doing.",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE": "Oppia's lessons, also known as explorations, provide more immersive experiences than static videos or text, helping users learn by doing.",
+  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Oppia's lessons, also known as explorations, provide more immersive experiences than static videos or text, helping users learn by doing.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_DESCRIPTION": "Oppia is a free, open-source learning platform full of interactive activities called 'explorations'.  Sadly, Oppia requires JavaScript to be enabled in your web browser in order to function properly and your web browser has JavaScript disabled.  If you need help enabling JavaScript, <a href=\"<[hrefUrl]>\">click here.</a>",
   "I18N_SPLASH_JAVASCRIPT_ERROR_THANKS": "Thank you.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_TITLE": "We Need JavaScript in Your Browser",
   "I18N_SPLASH_PAGE_TITLE": "Oppia: Teach, Learn, Explore",
   "I18N_SPLASH_SECOND_EXPLORATION_DESCRIPTION": "Explorations are simple to create. They are easily adapted based on individual student feedback and trends in the experience of your learners around the world.",
   "I18N_SPLASH_SITE_FEEDBACK": "Site Feedback",
-  "I18N_SPLASH_SUBTITLE": "Oppia makes it easy to create interactive explorations that educate and engage.",
-  "I18N_SPLASH_SUBTITLE_ALTERNATE": "Oppia makes it easy to create interactive lessons that educate and engage.",
+  "I18N_SPLASH_SUBTITLE": "Oppia makes it easy to create interactive lessons that educate and engage.",
   "I18N_SPLASH_THIRD_EXPLORATION_DESCRIPTION": "Oppia lets you create and share explorations on a wide range of subjects, limited only by your imagination.",
   "I18N_SPLASH_TITLE": "Think outside the books.",
   "I18N_TOPNAV_ABOUT": "About",
