+++ assets/i18n/es.json
@@ -265,16 +265,14 @@
   "I18N_SIGNUP_USERNAME_EXPLANATION": "Tu usuario se mostrará al lado de tus contribuciones.",
   "I18N_SIGNUP_WAIVER_OBJECTIVE": "La renuncia al requerimiento de atribución significa que, si alguien reutiliza ese trabajo, no están obligados a atribuirlo a los autores. Sin embargo, todas tus contribuciones individuales a las exploraciones estarán disponibles en el sitio, en el registro de cambios de la exploración. Se recomienda que quien reuse la exploración incluya un enlace a esta página, pero no es una obligación.",
   "I18N_SIGNUP_WHY_LICENSE": "¿Por qué CC-BY-SA?",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Las exploraciones son lecciones interactivas que proveen experiencias más envolventes que videos estáticos o texto, ayudando a los usuarios a aprender haciendo.",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE": "Las lecciones de Oppia, también conocidas como exploraciones, proveen experiencias más envolventes que videos estáticos o texto, ayudando a los usuarios a aprender haciendo.",
+  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Las lecciones de Oppia, también conocidas como exploraciones, proveen experiencias más envolventes que videos estáticos o texto, ayudando a los usuarios a aprender haciendo.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_DESCRIPTION": "Oppia es una plataforma de aprendizaje gratuita, libre y de código abierto Oppia llena de actividades interactivas llamadas exploraciones. Lamentablemente, Oppia requiere que JavaScript esté activado en tu navegador para funcionar apropiadamente y tu navegador ha desactivado JavaScript. Si necesitas ayuda activando JavaScript, <a href=\"<[hrefUrl]>\">haz click aquí.</a>",
   "I18N_SPLASH_JAVASCRIPT_ERROR_THANKS": "Thank you.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_TITLE": "Necesitamos JavaScript en tu navegador.",
   "I18N_SPLASH_PAGE_TITLE": "Oppia: Enseñar, Aprender, Explorar",
   "I18N_SPLASH_SECOND_EXPLORATION_DESCRIPTION": "Las exploraciones son simples de crear. Se adaptan fácilmente según comentarios individuales de los estudiantes y tendencias en la experiencia de los usuarios alrededor del mundo.",
   "I18N_SPLASH_SITE_FEEDBACK": "Comentarios al sitio",
-  "I18N_SPLASH_SUBTITLE": "Con Oppia es fácil crear exploraciones interactivas que educan y entretienen.",
-  "I18N_SPLASH_SUBTITLE_ALTERNATE": "Con Oppia es fácil crear lecciones interactivas que educan y entretienen.",
+  "I18N_SPLASH_SUBTITLE": "Con Oppia es fácil crear lecciones interactivas que educan y entretienen.",
   "I18N_SPLASH_THIRD_EXPLORATION_DESCRIPTION": "Oppia te permite crear y compartir exploraciones en una gran variedad de temas, limitado sólo por tu imaginación.",
   "I18N_SPLASH_TITLE": "Piensa más allá de los libros.",
   "I18N_TOPNAV_ABOUT": "Acerca de",
