+++ assets/i18n/id.json
@@ -265,16 +265,14 @@
   "I18N_SIGNUP_USERNAME_EXPLANATION": "Username anda akan ditampilkan di sebelah kontribusi anda.",
   "I18N_SIGNUP_WAIVER_OBJECTIVE": "Persyaratan pengabaian atribusi berarti jika seseorang menggunakan kembali materi ini, ia tidak diwajibkan menyatakan siapa penulis aslinya. Namun, semua kontribusi individu pada pembuatan eksplorasi akan tetap tercatat di catatan perubahan eksplorasi situs ini, dan orang-orang yang menggunakan kembali eksplorasi tersebut disarankan (tapi tidak diwajibkan) untuk menambahkan tautan ke halaman ini.",
   "I18N_SIGNUP_WHY_LICENSE": "Mengapa CC-BY-SA?",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Eksplorasi adalah pelajaran interaktif yang memberikan pengalaman lebih dalam dari video atau teks statis, membantu pengguna belajar sambil mempraktekkan.",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE": "Pelajaran di Oppia, juga dikenal sebagai eksplorasi, memberikan pengalaman lebih dalam dari video atau teks statis, membantu pengguna belajar sambil mempraktekkan.",
+  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Pelajaran di Oppia, juga dikenal sebagai eksplorasi, memberikan pengalaman lebih dalam dari video atau teks statis, membantu pengguna belajar sambil mempraktekkan.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_DESCRIPTION": "Oppia adalah platform pendidikan gratis dan open-source, penuh dengan aktivitas interaktif bernama 'eksplorasi'. Sayangnya, Oppia membutuhkan JavaScript diaktifkan dalam web browser anda agar dapat berfungsi dengan baik, namun JavaScript web browser anda nonaktif. Jika anda perlu bantuan mengaktifkan JavaScript, <a href=\"<[hrefUrl]>\">klik di sini.</a>",
   "I18N_SPLASH_JAVASCRIPT_ERROR_THANKS": "Terima kasih.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_TITLE": "Kami Membutuhkan JavaScript dalam Browser Anda",
   "I18N_SPLASH_PAGE_TITLE": "Oppia: Mengajar, Belajar, Menjelajah",
   "I18N_SPLASH_SECOND_EXPLORATION_DESCRIPTION": "Eksplorasi mudah dibuat. Sesuaikan materi dengan mudah berdasarkan feedback individual murid dan tren dari pengalaman semua pelajar anda dari seluruh dunia.",
   "I18N_SPLASH_SITE_FEEDBACK": "Feedback Situs",
-  "I18N_SPLASH_SUBTITLE": "Oppia mempermudah penciptaan eksplorasi interaktif yang mendidik dan menarik.",
-  "I18N_SPLASH_SUBTITLE_ALTERNATE": "Oppia mempermudah penciptaan pelajaran interaktif yang mendidik dan menarik.",
+  "I18N_SPLASH_SUBTITLE": "Oppia mempermudah penciptaan pelajaran interaktif yang mendidik dan menarik.",
   "I18N_SPLASH_THIRD_EXPLORATION_DESCRIPTION": "Dengan Oppia, ciptakan dan bagikan eksplorasimu dalam berbagai subjek, hanya imajinasimu batasnya.",
   "I18N_SPLASH_TITLE": "Berpikir di luar buku.",
   "I18N_TOPNAV_ABOUT": "Tentang",
