+++ assets/i18n/pt-br.json
@@ -265,16 +265,14 @@
   "I18N_SIGNUP_USERNAME_EXPLANATION": "Seu nome de usuário será mostrado ao lado de suas contribuições.",
   "I18N_SIGNUP_WAIVER_OBJECTIVE": "A renúncia da exigência de atribuição (BY) significa que, se alguém reutilizar esse trabalho, a atribuição aos autores não será obrigatória. No entanto, todas as suas contribuições individuais as explorações estarão disponíveis neste site no log de alterações da exploração, e as pessoas que reutilizam a exploração são encorajadas (mas não obrigadas) a incluir um link para esta página.",
   "I18N_SIGNUP_WHY_LICENSE": "Por que CC-BY-SA?",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Explorações são aulas interativas que proporcionam experiências mais envolventes do que vídeos estáticos ou textos, ajudando os usuários a aprender fazendo.",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE": "As aulas do Oppia, também conhecidas como explorações, proporcionam experiências mais envolventes do que vídeos estáticos ou textos, ajudando os usuários a aprender fazendo.",
+  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "As aulas do Oppia, também conhecidas como explorações, proporcionam experiências mais envolventes do que vídeos estáticos ou textos, ajudando os usuários a aprender fazendo.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_DESCRIPTION": "Oppia é uma plataforma de aprendizagem gratuita e de código aberto cheia de atividades interativas chamadas 'explorações'. Infelizmente, Oppia requer que o JavaScript esteja ativo em seu navegador web para funcionar corretamente e seu navegador está com o JavaScript desativado. Se precisar de ajuda para habilitar o JavaScript, <a href=\"<[hrefUrl]>\">clique aqui</a>.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_THANKS": "Muito obrigado.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_TITLE": "Precisamos de JavaScript em seu navegador",
   "I18N_SPLASH_PAGE_TITLE": "Oppia: Ensine, Aprenda, Explore",
   "I18N_SPLASH_SECOND_EXPLORATION_DESCRIPTION": "Explorações são fáceis de criar. Elas são facilmente adaptáveis com base no feedback de cada estudante e em tendências na experiência de estudantes do mundo todo.",
   "I18N_SPLASH_SITE_FEEDBACK": "Deixe sua opinião sobre o site",
-  "I18N_SPLASH_SUBTITLE": "Oppia facilita a criação de explorações interativas que educam e engajam.",
-  "I18N_SPLASH_SUBTITLE_ALTERNATE": "Oppia facilita a criação de aulas interativas que educam e engajam.",
+  "I18N_SPLASH_SUBTITLE": "Oppia facilita a criação de aulas interativas que educam e engajam.",
   "I18N_SPLASH_THIRD_EXPLORATION_DESCRIPTION": "Oppia lhe permite criar e compartilhar explorações sobre uma ampla gama de assuntos, limitado apenas pela sua imaginação.",
   "I18N_SPLASH_TITLE": "Pense de forma inovadora.",
   "I18N_TOPNAV_ABOUT": "Sobre",
