+++ assets/i18n/qqq.json
@@ -266,15 +266,13 @@
   "I18N_SIGNUP_WAIVER_OBJECTIVE":"Text displayed inside a information dialog widget in the signup page. - It explains the waiver of attribution. See I18N_SIGNUP_WHY_LICENSE.",
   "I18N_SIGNUP_WHY_LICENSE": "Text displayed inside a information dialog widget in the signup page. - It's the title of a dialog that contains additional information about why the creative commons licence was chosen for Oppia.",
   "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Paragraph describing Oppia. - Shown in the splash page next to an image.",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE": "Paragraph describing Oppia. - Shown in the splash page next to an image.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_DESCRIPTION": "Main content of an error page - The error is triggered when the user has disabled javascript in the browser. There is no space constraints for this paragraph and the priority is to clearly explain the user what caused the problem and how to solve it.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_THANKS": "Closing message of an error page - Thanks the user in a informal language.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_TITLE": "Title of an error page - The error is triggered when the user has disabled javascript in the browser.",
   "I18N_SPLASH_PAGE_TITLE": "Title displayed in the tab on the Browser in Oppia's landing page - Keep in three short words, the closes possible to the original meaning.",
   "I18N_SPLASH_SECOND_EXPLORATION_DESCRIPTION": "Paragraph describing Oppia showed in the splash page next to an image.",
   "I18N_SPLASH_SITE_FEEDBACK": "Text displayed in a floating tag at the bottom of the page - It links to the feedback page and should engage the user to give an opinion on the site. It should be only a couple words long.",
-  "I18N_SPLASH_SUBTITLE": "The second title shown in Oppia's landing page, below the main image - It should not exceed the two lines in length.",
-  "I18N_SPLASH_SUBTITLE_ALTERNATE": "An alternative second title shown in Oppia's landing page, below the main image - It should not exceed the two lines in length.",
+  "I18N_SPLASH_SUBTITLE": "The second title shown in Oppia's landing page, below the main image - It should not exceed two lines in length.",
   "I18N_SPLASH_THIRD_EXPLORATION_DESCRIPTION": "Paragraph describing Oppia showed in the splash page next to an image.",
   "I18N_SPLASH_TITLE": "The main title shown in Oppia's landing page - This is an English idiom, it can be changed for a local common saying with similar meaning. It should not exceed one line in length.",
   "I18N_TOPNAV_ABOUT": "Text displayed in the navigation bar. - When the user hovers over the button, a list with several information pages is displayed.",
