+++ assets/i18n/vi.json
@@ -265,16 +265,14 @@
   "I18N_SIGNUP_USERNAME_EXPLANATION": "Tên Tài Khoản Của Bạn Sẽ Xuất Hiện Kế Bên Những Đóng Góp Của Bạn.",
   "I18N_SIGNUP_WAIVER_OBJECTIVE": "Giấy ủy quyền (BY) có điều kiện, nếu người khác dùng tài liệu của bạn, họ không phải nêu tên bạn. Mọi tài liệu của bạn sẽ được đang trên Oppia và bất cư ai dùng sẽ có đề nghị từ Oppia nên để đường link tới trang tài liệu của bạn trên trang của họ.",
   "I18N_SIGNUP_WHY_LICENSE": "Tại sao lại sử dụng giấy phép CC-BY-SA?",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Những cuộc hành trình là các bài học có tính tương tác cao hơn so với video và bài giảng thông thường, với mục đích giúp các bạn hiểu bằng cách thực hành.",
-  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE": "Những bài học của Oppia, với cái tên gọi khác là nhửng cuộc hành trình, sẽ mang đến cho các bạn một trải nghiệm mới mẻ và thú vị hơn cách xem video hoặc đọc sách thông thường.",
+  "I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION": "Những bài học của Oppia, với cái tên gọi khác là nhửng cuộc hành trình, sẽ mang đến cho các bạn một trải nghiệm mới mẻ và thú vị hơn cách xem video hoặc đọc sách thông thường.",
   "I18N_SPLASH_JAVASCRIPT_ERROR_DESCRIPTION": "Oppia là một chưng trình phầm mềm tự do và miễn phí. Oppia cho phép bạn tạo ra những bài giảng có tính tương tác cao, gọi là 'cuộc hành trình'. Để chưng trình có thể hoạt đông, Oppia cần phải có Javascript trong trình duyệt web của bạn. Nếu bạn không biết cách cho phép Javascript hoạt động,  <a href=\"<[hrefUrl]>\">bấm vào đây.</a>",
   "I18N_SPLASH_JAVASCRIPT_ERROR_THANKS": "Cảm Ơn Bạn",
   "I18N_SPLASH_JAVASCRIPT_ERROR_TITLE": "Trình duyệt của bạn cần phải cài Javascript",
   "I18N_SPLASH_PAGE_TITLE": "Oppia: Dạy, Học và Khám Phá",
   "I18N_SPLASH_SECOND_EXPLORATION_DESCRIPTION": "Bạn có thể tạo ra những cuộc hành trình một cách đơn giản. Chúng có thể thay đổi dựa theo phản hồi của học sinh và các xu hướng dạy học trên thế gưới.",
   "I18N_SPLASH_SITE_FEEDBACK": "Gửi phản hồi",
-  "I18N_SPLASH_SUBTITLE": "Oppia cho phép bạn tạo những cuộc hành trình phù hợp với mọi lứa tuổi.",
-  "I18N_SPLASH_SUBTITLE_ALTERNATE": "Oppia cho phép bạn tạo các bài học phù hợp với mọi lứa tuổi.",
+  "I18N_SPLASH_SUBTITLE": "Oppia cho phép bạn tạo các bài học phù hợp với mọi lứa tuổi.",
   "I18N_SPLASH_THIRD_EXPLORATION_DESCRIPTION": "Oppia cho phép bạn tạo ra và chia sẻ các cuộc hành trình, và sự giới hạn duy nhất là chính trí tưởng tượng của ban.",
   "I18N_SPLASH_TITLE": "Suy Nghĩ Sáng Tạo.",
   "I18N_TOPNAV_ABOUT": "Giới Thiệu",
