+++ core/controllers/collection_viewer.py
@@ -18,7 +18,7 @@ from core.controllers import base
 from core.domain import collection_services
 from core.domain import config_domain
 from core.domain import rights_manager
-from core.domain import summary_services
+from core.domain import summary_tile_services
 from core.platform import models
 import feconf
 import utils
@@ -88,7 +88,7 @@ class CollectionDataHandler(base.BaseHandler):
 
         try:
             collection_dict = (
-                summary_services.get_learner_collection_dict_by_id(
+                summary_tile_services.get_learner_collection_dict_by_id(
                     collection_id, self.user_id,
                     allow_invalid_explorations=False))
         except Exception as e:
