+++ core/controllers/reader.py
@@ -36,7 +36,7 @@ from core.domain import rating_services
 from core.domain import recommendations_services
 from core.domain import rights_manager
 from core.domain import rte_component_registry
-from core.domain import summary_services
+from core.domain import summary_tile_services
 import feconf
 import utils
 
@@ -618,7 +618,7 @@ class RecommendationsHandler(base.BaseHandler):
 
         self.values.update({
             'summaries': (
-                summary_services.get_displayable_exp_summary_dicts_matching_ids(
+                summary_tile_services.get_displayable_exploration_summary_dicts_matching_ids( #pylint: disable=line-too-long
                     author_recommended_exp_ids + auto_recommended_exp_ids)),
         })
         self.render_json(self.values)
