+++ core/templates/dev/head/pages/Base.js
@@ -16,186 +16,23 @@
  * @fileoverview Oppia's base controller.
  */
 
-// TODO(sll): Get this to read from a common JSON file; it's replicated in
-// feconf.
-oppia.constant('CATEGORY_LIST', GLOBALS.ALL_CATEGORIES || []);
-
-// We use a slash because this character is forbidden in a state name.
-oppia.constant('PLACEHOLDER_OUTCOME_DEST', '/');
-
-oppia.constant('DEFAULT_RULE_NAME', 'Default');
-
-oppia.constant('CLASSIFIER_RULESPEC_STR', 'FuzzyMatches');
-
-// Feature still in development.
-// NOTE TO DEVELOPERS: This should be synchronized with the value in feconf.
-oppia.constant('ENABLE_STRING_CLASSIFIER', false);
-
-oppia.constant('DEFAULT_CLASSIFIER_RULE_SPEC', {
-  rule_type: 'FuzzyMatches',
-  inputs: {
-    training_data: []
-  }
-});
-
-oppia.constant('EVENT_HTML_CHANGED', 'htmlChanged');
-
-oppia.constant('PARAMETER_TYPES', {
-  REAL: 'Real',
-  UNICODE_STRING: 'UnicodeString'
-});
-
-oppia.constant('INTERACTION_DISPLAY_MODE_INLINE', 'inline');
-
-oppia.constant('OBJECT_EDITOR_URL_PREFIX', '/object_editor_template/');
-
-// The maximum number of nodes to show in a row of the state graph.
-oppia.constant('MAX_NODES_PER_ROW', 4);
-// The following variable must be at least 3. It represents the maximum length,
-// in characters, for the name of each node label in the state graph.
-oppia.constant('MAX_NODE_LABEL_LENGTH', 15);
-
-// If an $http request fails with the following error codes, a warning is
-// displayed.
-oppia.constant('FATAL_ERROR_CODES', [400, 401, 404, 500]);
-
-// Global utility methods.
 oppia.controller('Base', [
-  '$scope', '$http', '$rootScope', '$window', '$timeout', '$document', '$log',
-  'alertsService', 'LABEL_FOR_CLEARING_FOCUS', 'siteAnalyticsService',
-  'windowDimensionsService', 'UrlInterpolationService',
-  function(
-      $scope, $http, $rootScope, $window, $timeout, $document, $log,
-      alertsService, LABEL_FOR_CLEARING_FOCUS, siteAnalyticsService,
-      windowDimensionsService, UrlInterpolationService) {
-    $rootScope.DEV_MODE = GLOBALS.DEV_MODE;
-
+  '$scope', '$rootScope', '$document', 'alertsService', 'SidebarStatusService',
+  function($scope, $rootScope, $document, alertsService, SidebarStatusService) {
     $scope.alertsService = alertsService;
-    $scope.LABEL_FOR_CLEARING_FOCUS = LABEL_FOR_CLEARING_FOCUS;
 
+    $rootScope.DEV_MODE = GLOBALS.DEV_MODE;
     // If this is nonempty, the whole page goes into 'Loading...' mode.
     $rootScope.loadingMessage = '';
 
-    if (GLOBALS.userIsLoggedIn) {
-      // Show the number of unseen notifications in the navbar and page title,
-      // unless the user is already on the dashboard page.
-      $http.get('/notificationshandler').then(function(response) {
-        var data = response.data;
-        if ($window.location.pathname !== '/') {
-          $scope.numUnseenNotifications = data.num_unseen_notifications;
-          if ($scope.numUnseenNotifications > 0) {
-            $window.document.title = (
-              '(' + $scope.numUnseenNotifications + ') ' +
-              $window.document.title);
-          }
-        }
-      });
-    }
-
-    $scope.getStaticImageUrl = UrlInterpolationService.getStaticImageUrl;
-
-    // This method is here because the trigger for the tutorial is in the site
-    // navbar. It broadcasts an event to tell the exploration editor to open the
-    // editor tutorial.
-    $scope.openEditorTutorial = function() {
-      $scope.$broadcast('openEditorTutorial');
-    };
-
-    // The following methods and listeners relate to the global navigation
-    // sidebar.
-    $scope.openSidebar = function() {
-      if (!$scope.sidebarIsShown) {
-        $scope.sidebarIsShown = true;
-        $scope.pendingSidebarClick = true;
-      } else {
-        $scope.sidebarIsShown = false;
-        $scope.pendingSidebarClick = false;
-      }
-    };
+    $scope.isSidebarShown = SidebarStatusService.isSidebarShown;
+    $scope.openSidebarOnSwipe = SidebarStatusService.openSidebar;
+    $scope.closeSidebarOnSwipe = SidebarStatusService.closeSidebar;
 
     // TODO(sll): use 'touchstart' for mobile.
     $document.on('click', function() {
-      if (!$scope.pendingSidebarClick) {
-        $scope.sidebarIsShown = false;
-      } else {
-        $scope.pendingSidebarClick = false;
-      }
+      SidebarStatusService.onDocumentClick();
       $scope.$apply();
     });
-
-    $scope.profileDropdownIsActive = false;
-
-    $scope.onMouseoverProfilePictureOrDropdown = function(evt) {
-      angular.element(evt.currentTarget).parent().addClass('open');
-      $scope.profileDropdownIsActive = true;
-    };
-
-    $scope.onMouseoutProfilePictureOrDropdown = function(evt) {
-      angular.element(evt.currentTarget).parent().removeClass('open');
-      $scope.profileDropdownIsActive = false;
-    };
-
-    $scope.onMouseoverDropdownMenu = function(evt) {
-      angular.element(evt.currentTarget).parent().addClass('open');
-    };
-
-    $scope.onMouseoutDropdownMenu = function(evt) {
-      angular.element(evt.currentTarget).parent().removeClass('open');
-    };
-
-    $scope.onLoginButtonClicked = function(loginUrl) {
-      siteAnalyticsService.registerStartLoginEvent('loginButton');
-      $timeout(function() {
-        $window.location = loginUrl;
-      }, 150);
-      return false;
-    };
-
-    var doesNavbarHaveSearchBar = function() {
-      return (
-        $window.location.pathname.indexOf('/search') === 0 ||
-        $window.location.pathname.indexOf('/library') === 0);
-    };
-
-    var navbarCutoffWidthPx = doesNavbarHaveSearchBar() ? 1171 : 800;
-
-    $scope.windowIsNarrow = (
-      windowDimensionsService.getWidth() <= navbarCutoffWidthPx);
-
-    $scope.openSidebarOnSwipe = function() {
-      if ($scope.windowIsNarrow) {
-        if (!$scope.sidebarIsShown) {
-          $scope.sidebarIsShown = true;
-          $scope.pendingSidebarClick = true;
-        }
-      }
-    };
-
-    $scope.closeSidebarOnSwipe = function() {
-      if ($scope.windowIsNarrow) {
-        if ($scope.sidebarIsShown) {
-          $scope.sidebarIsShown = false;
-          $scope.pendingSidebarClick = false;
-        }
-      }
-    };
-
-    windowDimensionsService.registerOnResizeHook(function() {
-      $scope.windowIsNarrow = (
-        windowDimensionsService.getWidth() <= navbarCutoffWidthPx);
-      $scope.$apply();
-
-      // If the window is now wide, and the sidebar is still open, close it.
-      if (!$scope.windowIsNarrow) {
-        $scope.sidebarIsShown = false;
-      }
-    });
-
-    $scope.pageHasLoaded = false;
-    $scope.pendingSidebarClick = false;
-    $scope.sidebarIsShown = false;
-    $timeout(function() {
-      $scope.pageHasLoaded = true;
-    }, 500);
   }
 ]);
