+++ core/templates/dev/head/pages/splash/splash.html
@@ -153,7 +153,7 @@
       </div>
 
       <div class="oppia-splash-h2-container">
-        <h2 class="oppia-splash-h2" translate="I18N_SPLASH_SUBTITLE_ALTERNATE" style="padding-bottom: 30px; font-size: 1.6em; padding-top: 100px; max-width: 700px;"></h2>
+        <h2 class="oppia-splash-h2" translate="I18N_SPLASH_SUBTITLE" style="padding-bottom: 30px; font-size: 1.6em; padding-top: 100px; max-width: 700px;"></h2>
       </div>
     </div>
 
@@ -164,7 +164,7 @@
             <img ng-src="<[getStaticImageUrl('/splash/bullet1icon.svg')]>" class="oppia-splash-overlapping-image-1">
           </div>
           <div class="oppia-splash-bullet-block oppia-splash-block-right-text"
-               translate="I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE">
+               translate="I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION">
           </div>
         </div>
