+++ core/templates/dev/head/pages/splash/splash_ah2.html
@@ -172,7 +172,7 @@
             <img ng-src="<[getStaticImageUrl('/splash/bullet1icon.svg')]>" class="oppia-splash-overlapping-image-1">
           </div>
           <div class="oppia-splash-bullet-block oppia-splash-block-right-text"
-               translate="I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION_ALTERNATE">
+               translate="I18N_SPLASH_FIRST_EXPLORATION_DESCRIPTION">
           </div>
         </div>
